Old code cleanup (removal)

+++ Incoming.py
@@ -5,7 +5,7 @@ import webapp2
 
 from django.utils import simplejson as json
 
-from MasterHandler import MasterHandler
+from MasterHandler import MasterHandler, AjaxError
 from users.UserDataModels import UserProfile, Psinque, UserEmail
 from users.UserManagement import getPublicGroup, getDisplayNameFromPsinque
 from users.Email import notifyPendingPsinque
@@ -16,11 +16,14 @@ class IncomingHandler(MasterHandler):
 
     def get(self, actionName):
         
-        if MasterHandler.safeGuard(self):   # perform action only if user is logged in
+        if MasterHandler.safeGuard(self):
             
             actionFunction = getattr(self, actionName)
-            actionFunction()
-
+            
+            try:
+                actionFunction()
+            except AjaxError as e:
+                self.sendJsonError(e.value)
 
     def view(self):
         
@@ -103,8 +106,9 @@ class IncomingHandler(MasterHandler):
 
     def removepsinque(self):
         
-        psinque = Psinque.get(self.request.get('key'))
-        psinque.delete()
+        psinque = Psinque.get(self.checkGetParameter('key'))
+        if not psinque is None:
+            psinque.delete()
         self.redirect("/incoming")
         
 #-----------------------------------------------------------------------------
